pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "whispercpp-local-poc"
version = "0.1.0"
description = "Mac-local whisper.cpp proof of concept for dysarthric speech"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.0.0",
  "pyobjc-framework-Cocoa>=11.0",
  "sounddevice>=0.5.0",
  "soundfile>=0.13.0",
  "webrtcvad-wheels>=2.0.14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
